+++ src/python/deepgnn/graph_engine/snark/decoders.py
@@ -5,7 +5,7 @@
 import abc
 import json
 import csv
-from typing import TypeVar, Iterator, Tuple
+from typing import TypeVar, Iterator, Tuple, Optional, List
 import numpy as np
 
 
@@ -41,6 +41,186 @@ class Decoder(abc.ABC):
 
 
 DecoderType = TypeVar("DecoderType", bound=Decoder)
+
+
+class LinearDecoder(Decoder):
+    """Convert the text line into node object.
+
+    Linear Format:
+        ```
+        <node info>
+        <edge_1_info>
+        <edge_2_info>
+        ...
+        ```
+        node_info: node_id -1 node_type node_weight <features>
+        edge_info: src dst edge_type edge_weight <features>
+        features[dense]: dtype_name length v1 v2 ... dtype_name2 length2 v1 v2 ...
+        features[sparse]: dtype_name values.size,coords.shape[1] c1 c2 ... v1 v2 ...
+        features[sparse]: dtype_name values.size,0 c1 c2 ... v1 v2 ...
+        * Nodes must be sorted by node_id, edges sorted by src and then dst.
+
+    Linear Format Example
+        A graph with 2 nodes {0, 1} each with type = 1, weight = .5 and
+        feature vectors [1, 1, 1] dtype=int32 and [1.1, 1.1] dtype=float32.
+        Edges: {0 -> 1, 1 -> 0} both with type = 0, weight = .5 and a sparse feature
+        vector (coords=[0, 4, 10], values=[1, 1, 1] dtype=uint8).
+        ```
+        0 -1 1 .5 int32 3 1 1 1 float32 2 1.1 1.1
+        0 1 0 .5 uint8 3,0 0 4 10 1 1 1
+        1 -1 1 .5 int32 3 1 1 1 float32 2 1.1 1.1
+        1 0 0 .5 uint8 3,0 0 4 10 1 1 1
+        ```
+
+    Init Parameters:
+        the following keyword arguments can be added when creating the decoder.
+        default_node_type: int Type of all nodes, if set do not add node type to any nodes.
+        default_node_weight: int Weight of all nodes, if set do not add node weight to any nodes.
+        default_node_feature_types: ["dtype" or None, ...] Dtype of each feature vector.
+        default_node_feature_lens: [[int, ...] or None, ...] Length value for each feature vector.
+        default_edge_type: int Same as node except for all edges.
+        default_edge_weight: int Same as node except for all edges.
+        default_edge_feature_types: ["dtype" or None, ...] Dtype of each feature vector.
+        default_edge_feature_lens: [[int, ...] or None, ...] Length value for each feature vector.
+        ```
+        LinearDecoder(
+            default_node_type=1,
+            default_node_weight=.5,
+            default_node_feature_types=["int32", "float32"],
+            default_node_feature_lens=[[3],[2]],
+            default_edge_type=0,
+            default_edge_weight=.5,
+            default_edge_feature_types=["uint8"],
+                default_edge_feature_lens=[[3, 0]],
+        )
+        ```
+        graph.linear
+        ```
+        0 -1 1 1 1 1.1 1.1
+        0 1 0 4 10 1 1 1
+        1 -1 1 1 1 1.1 1.1
+        1 0 0 4 10 1 1 1
+        ```
+    """
+
+    def __init__(
+        self,
+        default_node_type: Optional[int] = None,
+        default_edge_type: Optional[int] = None,
+        default_node_weight: Optional[float] = None,
+        default_edge_weight: Optional[float] = None,
+        default_node_feature_types: Optional[List[Optional[str]]] = None,
+        default_edge_feature_types: Optional[List[Optional[str]]] = None,
+        default_node_feature_lens: Optional[List[Optional[List[int]]]] = None,
+        default_edge_feature_lens: Optional[List[Optional[List[int]]]] = None,
+    ):
+        """Initialize the Decoder."""
+        super().__init__()
+        self.node_type = default_node_type
+        self.node_weight = default_node_weight
+        self.node_feature_types = (
+            default_node_feature_types if default_node_feature_types is not None else []
+        )
+        self.node_feature_lens = (
+            default_node_feature_lens if default_node_feature_lens is not None else []
+        )
+        self.n_node_feature = len(self.node_feature_types)
+
+        self.edge_type = default_edge_type
+        self.edge_weight = default_edge_weight
+        self.edge_feature_types = (
+            default_edge_feature_types if default_edge_feature_types is not None else []
+        )
+        self.edge_feature_lens = (
+            default_edge_feature_lens if default_edge_feature_lens is not None else []
+        )
+        self.n_edge_feature = len(self.edge_feature_types)
+
+    def decode(self, line: str) -> Iterator[Tuple[int, int, int, float, list]]:
+        """Convert text line into a node and edge iterator.
+
+        This is a generator that yields a node then its outgoing edges in order.
+        A node and its outgoing edges can be on different partitions.
+        Yield format is (node_id/src, -1/dst, type, weight, features).
+        Features being a list of dense features as ndarrays and sparse features as 2 tuples, coordinates and values.
+        """
+        if line == "":
+            return []
+
+        data = iter(line.split())
+
+        src, dst = int(next(data)), int(next(data))
+        if dst == -1:
+            typ, weight = self.node_type, self.node_weight
+            item_feature_types, item_feature_lens, default_feature_len = (
+                self.node_feature_types,
+                self.node_feature_lens,
+                self.n_node_feature,
+            )
+        else:
+            typ, weight = self.edge_type, self.edge_weight
+            item_feature_types, item_feature_lens, default_feature_len = (
+                self.edge_feature_types,
+                self.edge_feature_lens,
+                self.n_edge_feature,
+            )
+        if typ is None:
+            typ = int(next(data))
+        if weight is None:
+            weight = float(next(data))
+
+        features = []
+        n_features = 0
+        while True:
+            if default_feature_len > n_features:
+                key = item_feature_types[n_features]
+                length = item_feature_lens[n_features]
+            else:
+                key, length = None, None  # type: ignore
+
+            try:
+                if key is None:
+                    key = next(data)
+                    try:
+                        int(key)
+                        raise ValueError("Expected feature vector key is str.")
+                    except ValueError:
+                        pass
+                if length is None:
+                    length = list(map(int, next(data).split(",")))
+            except StopIteration:
+                break
+
+            if len(length) > 1:
+                values_len, coords_dim = length
+                if not values_len:
+                    value = None
+                else:
+                    value = (
+                        np.fromiter(
+                            data, count=values_len * max(coords_dim, 1), dtype=np.int64
+                        ).reshape(
+                            (values_len, coords_dim) if coords_dim else values_len
+                        ),
+                        np.fromiter(
+                            data,
+                            count=values_len,
+                            dtype=key,
+                        ),
+                    )
+            else:
+                length_single = length[0]
+                if not length_single:
+                    value = None
+                elif length_single == 1 and key == "binary_feature":
+                    value = next(data)  # type: ignore
+                else:
+                    value = np.fromiter(data, count=length_single, dtype=key)  # type: ignore
+
+            features.append(value)
+            n_features += 1
+
+        yield src, dst, typ, weight, features
 
 
 class JsonDecoder(Decoder):